refactor: 파일명 Join1Context로 수정 #28
