+++ src/app/(mood)/join/funnels/step1/components/Keywords.tsx
@@ -3,7 +3,7 @@
 import { KEYWORD_LIST } from '@/constants/keywords';
 import { cn } from '@/utils/cn';
 import { type UseFormReturn, useWatch } from 'react-hook-form';
-import { type Join1ContextValue } from '../../../components/MoodContext';
+import { type Join1ContextValue } from '../../../components/Join1Context';
 import PlusIcon from '@public/svg/plus-24.svg';
 import CheckIcon from '@public/svg/check-24.svg';
