+++ src/app/(mood)/join/funnels/step1/components/MoodBoard.tsx
@@ -3,7 +3,7 @@
 import { useTheme } from 'next-themes';
 import ToggleButton from './ToggleButton';
 import { useIsMounted } from '@/hooks';
-import { useJoin1Context } from '../../../components/MoodContext';
+import { useJoin1Context } from '../../../components/Join1Context';
 import { useWatch } from 'react-hook-form';
 import { motion } from 'framer-motion';
 import { fadeInOut } from '@/constants';
