+++ src/app/(mood)/join/funnels/step1/components/MyKeywordPopup.tsx
@@ -3,7 +3,7 @@ import { AnimatePortal } from '@/components/Modal';
 import PopupContainer from '@/components/Modal/PopupContainer';
 import Keywords from './Keywords';
 import { useWatch, type UseFormReturn } from 'react-hook-form';
-import type { Join1ContextValue } from '../../../components/MoodContext';
+import type { Join1ContextValue } from '../../../components/Join1Context';
 import Spacing from '@/components/Spacing';
 import Button from '@/components/Button/Button';
 import DeleteIcon from '@public/svg/delete-24.svg';
