+++ src/app/(mood)/join/funnels/step1/components/PostCodePopup.tsx
@@ -2,7 +2,7 @@
 
 import { Portal } from '@/components/Modal';
 import DaumPostcodeEmbed, { type Address } from 'react-daum-postcode';
-import type { Join1ContextValue } from '../../../components/MoodContext';
+import type { Join1ContextValue } from '../../../components/Join1Context';
 import { type UseFormReturn } from 'react-hook-form';
 import PopupHeader from '../../../components/PopupHeader';
