+++ src/app/(mood)/join/funnels/step1/components/RegionSection.tsx
@@ -1,6 +1,6 @@
 import Spacing from '@/components/Spacing';
 import { cn } from '@/utils/cn';
-import { useJoin1Context } from '../../../components/MoodContext';
+import { useJoin1Context } from '../../../components/Join1Context';
 import RegionIcon from '@/assets/Region';
 import RightArrowIcon from '@/assets/RightArrow';
 import useOverlay from '@/hooks/useOverlay';
